:recycle: refactor: seperate store & array util
